CR corrections

+++ sdk/python/feast/type_map.py
@@ -488,10 +488,10 @@ def bq_to_feast_value_type(bq_type_as_str):
     type_map: Dict[ValueType, Union[str, Dict[str, Any]]] = {
         "DATETIME": ValueType.STRING,  # Update to ValueType.UNIX_TIMESTAMP once #1520 lands.
         "TIMESTAMP": ValueType.STRING,  # Update to ValueType.UNIX_TIMESTAMP once #1520 lands.
-        "INTEGER": ValueType.INT32,
+        "INTEGER": ValueType.INT64,
         "INT64": ValueType.INT64,
         "STRING": ValueType.STRING,
-        "FLOAT": ValueType.FLOAT,
+        "FLOAT": ValueType.DOUBLE,
         "FLOAT64": ValueType.DOUBLE,
         "BYTES": ValueType.BYTES,
         "BOOL": ValueType.BOOL,
